Refactoring BLAS Kernel Helper for Backend Efficiency
This commit removes blas_kernel_helper.h and moves the operations to cpu_backend.
This enables efficient management of the function across different backends.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped
