+++ nntrainer/tensor/cpu_backend/fallback/fallback.cpp
@@ -173,6 +173,11 @@ void copy_u16(const unsigned int N, const uint16_t *X, uint16_t *Y) {
   __fallback_copy_u16(N, X, Y);
 }
 
+void convert_q4_0x8_shuffle_dispatch(const void *src, uint16_t *d_out,
+                                     uint8_t *qs_out, int N, int K) {
+  __fallback_convert_q4_0x8_shuffle_dispatch(src, d_out, qs_out, N, K);
+}
+
 void calc_trigonometric_vals_dup(unsigned int N_half, float *angle, float *cos_,
                                  float *sin_, unsigned int alpha) {
   __fallback_calc_trigonometric_vals_dup(N_half, angle, cos_, sin_, alpha);
