+++ nntrainer/tensor/cpu_backend/fallback/fallback.h
@@ -412,6 +412,12 @@ void transpose_matrix(const unsigned int M, const unsigned int N,
 #endif
 
 /**
+ * @copydoc convert_q4_0x8_shuffle_dispatch in cpu_backend.h
+ */
+void convert_q4_0x8_shuffle_dispatch(const void *src, uint16_t *d_out,
+                                     uint8_t *qs_out, int N, int K);
+
+/**
  * @brief Get half-sized angles, transform them into each cos, sin, and scopy in
  * the same vector : cos_ = cos(freq).extend(cos(freq)), sin_ =
  * sin(freq).extend(sin_(req))
