+++ nntrainer/tensor/cpu_backend/fallback/fallback_internal.h
@@ -424,6 +424,14 @@ _FP16 __fallback_max(const unsigned int N, _FP16 *X);
  */
 void __fallback_softmax(const unsigned int N, _FP16 *X, _FP16 *Y);
 #endif
+
+/**
+ * @copydoc convert_q4_0x8_shuffle_dispatch in cpu_backend.h
+ */
+void __fallback_convert_q4_0x8_shuffle_dispatch(const void *src,
+                                                uint16_t *d_out,
+                                                uint8_t *qs_out, int N, int K);
+
 /**
  * @brief Get half-sized angles, transform them into each cos, sin, and scopy in
  * the same vector : cos_ = cos(freq).extend(cos(freq)), sin_ =
